+++ x/epoching/genesis_test.go
@@ -28,9 +28,9 @@ func TestInitGenesis(t *testing.T) {
 
 	genesisState := types.GenesisState{
 		Params: types.Params{
-			EpochInterval:  100,
-			EnqueueGasFees: types.DefaultEnqueueGasFees,
-			MinAmount:      types.DefaultMinAmount,
+			EpochInterval: 100,
+			ExecuteGas:    types.DefaultExecuteGas,
+			MinAmount:     types.DefaultMinAmount,
 		},
 	}