+++ x/epoching/keeper/delegation_pool.go
@@ -9,9 +9,8 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
-// LockUserFunds locks user funds in delegate pool during message enqueue
-
-func (k Keeper) LockFunds(ctx sdk.Context, msg *types.QueuedMessage) error {
+// LockFunds locks user funds in delegate pool until the message is executed
+func (k Keeper) LockFundsForDelegateMsgs(ctx sdk.Context, msg *types.QueuedMessage) error {
 	msgId := hex.EncodeToString(msg.MsgId)
 	switch wrappedMsg := msg.Msg.(type) {
 	case *types.QueuedMessage_MsgDelegate:
@@ -30,7 +29,8 @@ func (k Keeper) LockFunds(ctx sdk.Context, msg *types.QueuedMessage) error {
 	}
 }
 
-func (k Keeper) UnLockFunds(ctx sdk.Context, msg *types.QueuedMessage) error {
+// UnLockFunds unlocks user funds from delegate pool when the message is executed
+func (k Keeper) UnlockFundsForDelegateMsgs(ctx sdk.Context, msg *types.QueuedMessage) error {
 	msgId := hex.EncodeToString(msg.MsgId)
 	switch wrappedMsg := msg.Msg.(type) {
 	case *types.QueuedMessage_MsgDelegate:
@@ -49,13 +49,14 @@ func (k Keeper) UnLockFunds(ctx sdk.Context, msg *types.QueuedMessage) error {
 	}
 }
 
+// lockDelegateFunds transfers the delegation amount from the delegator's account to the delegation pool module account
 func (k Keeper) lockDelegateFunds(ctx sdk.Context, msgId string, msg *types.MsgWrappedDelegate) error {
 	delegatorAddr, err := sdk.AccAddressFromBech32(msg.Msg.DelegatorAddress)
 	if err != nil {
 		return err
 	}
 
-	// Transfer funds from user to delegate pool
+	// Transfer funds from delegator to delegation pool
 	coins := sdk.NewCoins(msg.Msg.Amount)
 	if err := k.bk.SendCoinsFromAccountToModule(ctx, delegatorAddr, types.DelegatePoolModuleName, coins); err != nil {
 		return errorsmod.Wrapf(err, "failed to lock delegate funds for msg %s", msgId)
@@ -64,6 +65,7 @@ func (k Keeper) lockDelegateFunds(ctx sdk.Context, msgId string, msg *types.MsgW
 	return nil
 }
 
+// lockValidatorFunds transfers the delegation amount from the validator's account to the delegation pool module account
 func (k Keeper) lockValidatorFunds(ctx sdk.Context, msgId string, msg *checkpointingtypes.MsgWrappedCreateValidator) error {
 	validatorAddr, err := sdk.ValAddressFromBech32(msg.MsgCreateValidator.ValidatorAddress)
 	if err != nil {
@@ -71,7 +73,7 @@ func (k Keeper) lockValidatorFunds(ctx sdk.Context, msgId string, msg *checkpoin
 	}
 	valAddr := sdk.AccAddress(validatorAddr)
 
-	// Transfer funds from user to delegate pool
+	// Transfer funds from validator to delegation pool
 	coins := sdk.NewCoins(msg.MsgCreateValidator.Value)
 	if err := k.bk.SendCoinsFromAccountToModule(ctx, valAddr, types.DelegatePoolModuleName, coins); err != nil {
 		return errorsmod.Wrapf(err, "failed to lock delegate funds for msg %s", msgId)
@@ -80,33 +82,33 @@ func (k Keeper) lockValidatorFunds(ctx sdk.Context, msgId string, msg *checkpoin
 	return nil
 }
 
+// unlockDelegateFunds transfers the delegation amount from the delegation pool module account to the delegator's account
 func (k Keeper) unlockDelegateFunds(ctx sdk.Context, msgId string, msg *types.MsgWrappedDelegate) error {
 	delegatorAddr, err := sdk.AccAddressFromBech32(msg.Msg.DelegatorAddress)
 	if err != nil {
 		return err
 	}
-	// Transfer funds back from delegate pool to user
+	// Transfer funds back from delegation pool to delegator
 	coins := sdk.NewCoins(msg.Msg.Amount)
 	if err := k.bk.SendCoinsFromModuleToAccount(ctx, types.DelegatePoolModuleName, delegatorAddr, coins); err != nil {
 		return errorsmod.Wrapf(err, "failed to unlock delegate funds for msg %s", msgId)
 	}
 
-	// Remove locked fund record
 	return nil
 }
 
+// unlockValidatorFunds transfers the delegation amount from the delegation pool module account to the validator's account
 func (k Keeper) unlockValidatorFunds(ctx sdk.Context, msgId string, msg *checkpointingtypes.MsgWrappedCreateValidator) error {
 	validatorAddr, err := sdk.ValAddressFromBech32(msg.MsgCreateValidator.ValidatorAddress)
 	if err != nil {
 		return err
 	}
 	valAddr := sdk.AccAddress(validatorAddr)
-	// Transfer funds back from delegate pool to user
+	// Transfer funds back from delegation pool to validator
 	coins := sdk.NewCoins(msg.MsgCreateValidator.Value)
 	if err := k.bk.SendCoinsFromModuleToAccount(ctx, types.DelegatePoolModuleName, valAddr, coins); err != nil {
 		return errorsmod.Wrapf(err, "failed to unlock delegate funds for msg %s", msgId)
 	}
 
-	// Remove locked fund record
 	return nil
 }